requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1615404881292/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1614360992924/work
antlr4-python3-runtime==4.8
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
autopep8==1.5.4
backcall==0.2.0
bleach==3.2.1
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1611555765219/work
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246945912/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1602255302199/work
click==7.1.2
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1616851476134/work
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dicom2nifti==2.2.12
einops==0.3.0
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.8.4
fsspec==0.8.4
future==0.18.2
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1616709862146/work
google-auth-oauthlib==0.4.2
grpcio==1.34.0
hydra-core==1.0.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
imageio==2.9.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1616943212245/work
importlib-resources==3.3.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1614595805172/work
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
multidict @ file:///tmp/build/80754af9/multidict_1607367757617/work
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nibabel==3.2.1
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
omegaconf==2.0.5
opencv-python==4.5.1.48
openpyxl==3.0.6
packaging==20.7
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydicom==2.1.2
pyflakes==2.2.0
Pygments==2.7.3
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1610910308735/work
pylint==2.6.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291447907/work
python-dateutil==2.8.1
pytorch-lightning==1.0.8
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1595492159598/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1614171254180/work
scikit-image==0.18.0
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
SimpleITK==2.0.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sklearn==0.0
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1610699261066/work/tensorboard-2.4.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1611075653546/work/tensorboard_plugin_wit-1.8.0-py3-none-any.whl
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.12.8
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.54.0
traitlets==5.0.5
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1615828766818/work
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1605429457708/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
